feat(quickstarts): automate kcat quickstart

+++ src/main/java/io/managed/services/test/cli/AsyncProcess.java
@@ -1,32 +1,80 @@
 package io.managed.services.test.cli;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import lombok.extern.log4j.Log4j2;
+import org.apache.commons.io.input.TeeInputStream;
+import org.apache.commons.io.output.ByteArrayOutputStream;
+
 import java.io.BufferedReader;
+import java.io.BufferedWriter;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.io.OutputStream;
+import java.io.OutputStreamWriter;
+import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 
-import static io.managed.services.test.cli.ProcessUtils.buffer;
 import static lombok.Lombok.sneakyThrow;
 
-class AsyncProcess {
+@Log4j2
+public class AsyncProcess {
 
     private final Process process;
     private final ProcessHandle.Info info;
 
-    AsyncProcess(Process process) {
+    /**
+     * Joint output from stdout and stderr similar to what will be printed on a terminal
+     */
+    private final ByteArrayOutputStream joint = new ByteArrayOutputStream();
+
+    private final ByteArrayOutputStream stdout = new ByteArrayOutputStream();
+    private final ByteArrayOutputStream stderr = new ByteArrayOutputStream();
+
+    private final InputStream liveStdout;
+    private final InputStream liveStderr;
+    private final OutputStream liveStdin;
+
+    private final Exception cause;
+
+    private CompletableFuture<Process> future;
+
+    private boolean destroyed;
+
+    public AsyncProcess(Process process) {
         this.info = process.info();
         this.process = process;
+
+        this.liveStdout = new TeeInputStream(new TeeInputStream(process.getInputStream(), joint), stdout);
+        this.liveStderr = new TeeInputStream(new TeeInputStream(process.getErrorStream(), joint), stderr);
+
+        this.liveStdin = process.getOutputStream();
+
+        this.cause = new Exception();
     }
 
     public CompletableFuture<Process> future(Duration timeout) {
-        var cause = new Exception();
-        return process.onExit()
+        if (future != null) {
+            return future;
+        }
+
+        future = process.onExit()
 
             // set a timeout for the process
             .orTimeout(timeout.toMillis(), TimeUnit.MILLISECONDS)
 
             // handle the process result
             .handle((p, t) -> {
+
+                if (!destroyed) {
+                    // if the process has been destroyed the stderr and stdout are closed
+                    readAll();
+                }
 
                 if (process.isAlive()) {
                     process.destroyForcibly();
@@ -37,18 +85,109 @@ class AsyncProcess {
                 }
 
                 if (p.exitValue() != 0) {
-                    throw sneakyThrow(new ProcessException(p, info, cause));
+                    throw sneakyThrow(processException(p.exitValue(), cause));
                 }
                 return p;
             });
+        return future;
+    }
+
+    public AsyncProcess sync(Duration timeout) throws ProcessException {
+
+        try {
+            future(timeout).get();
+            return this;
+
+        } catch (ExecutionException e) {
+            var cause = e.getCause();
+            if (cause instanceof ProcessException) {
+                throw (ProcessException) cause;
+            }
+            throw sneakyThrow(cause);
+        } catch (InterruptedException e) {
+            throw sneakyThrow(e);
+        }
+    }
+
+    public ProcessException processException(Exception cause) {
+        return processException(process.exitValue(), cause);
+    }
+
+    private ProcessException processException(int exitCode, Exception cause) {
+        return new ProcessException(
+            info.commandLine().orElse(""),
+            exitCode,
+            stdoutAsString(),
+            stderrAsString(),
+            outputAsString(),
+            cause);
+    }
+
+    public boolean isDead() {
+        return !process.isAlive();
+    }
+
+    public BufferedWriter stdin() {
+        return new BufferedWriter(new OutputStreamWriter(liveStdin));
     }
 
     public BufferedReader stdout() {
-        return buffer(process.getInputStream());
+        return new BufferedReader(new InputStreamReader(liveStdout));
     }
 
-    @SuppressWarnings("unused")
-    public BufferedReader stderr() {
-        return buffer(process.getErrorStream());
+    /**
+     * @return the full output written to stdout till now
+     */
+    public String stdoutAsString() {
+        return stdout.toString(StandardCharsets.UTF_8);
+    }
+
+    /**
+     * @return the full output written to stderr till now
+     */
+    public String stderrAsString() {
+        return stderr.toString(StandardCharsets.UTF_8);
+    }
+
+    public String outputAsString() {
+        return joint.toString(StandardCharsets.UTF_8);
+    }
+
+    public <T> T asJson(Class<T> c) {
+        try {
+            return new ObjectMapper()
+                .registerModule(new JavaTimeModule())
+                .readValue(stdoutAsString(), c);
+        } catch (JsonProcessingException e) {
+            throw sneakyThrow(e);
+        }
+    }
+
+    public void readAll() {
+        readAll(liveStdout);
+        readAll(liveStderr);
+    }
+
+    private static void readAll(InputStream stream) {
+        try {
+            stream.readNBytes(stream.available());
+        } catch (IOException e) {
+            log.error("failed to read all remaining bytes form the input stream:", e);
+        }
+    }
+
+    public void destroy() {
+
+        if (isDead()) {
+            return;
+        }
+
+        // save stdin and stdout before destroy the process
+        readAll();
+
+        this.destroyed = true;
+
+        // destroy the process
+        process.destroy();
     }
 }
